poprawki przy usuwaniu oddzialow

+++ bitwa/pole/PoleBitwy.cpp
@@ -69,10 +69,10 @@ Oddzial *PoleBitwy::znajdzPrzeciwnika(Oddzial& atakujacy, uint8_t nrPrzeciwnika)
         return poleGry_.at(nrPrzeciwnika).at(0).at(nrKolumny)->zwrocOddzial();
     }
     unsigned int zasieg=atakujacy.getZasieg_();
-    int odleglosc=1;
-    while(odleglosc<=zasieg)
+    unsigned int odleglosc = 1;
+    while(odleglosc <= zasieg)
     {
-        if(static_cast<int>(nrKolumny) - odleglosc >= 0
+        if(static_cast<int>(nrKolumny) - static_cast<int>(odleglosc) >= 0
            && poleGry_.at(nrPrzeciwnika).at(0).at(nrKolumny-odleglosc)->zwrocOddzial()!= nullptr)
         {
             return poleGry_.at(nrPrzeciwnika).at(0).at(nrKolumny-odleglosc)->zwrocOddzial();
